pluggable almost working
override weak function attachFunctionsToUSB

\#include "HID.h"
HID.begin() to plug it

+++ hardware/arduino/avr/cores/arduino/USBAPI.h
@@ -27,7 +27,7 @@
 #include <util/delay.h>
 
 typedef unsigned char u8;
-typedef unsigned int u16;
+typedef unsigned short u16;
 typedef unsigned long u32;
 
 #include "Arduino.h"
